+++ src/lexer/stream.h
@@ -12,11 +12,11 @@ class LocatableStream {
         , m_filename(filename) {}
 
     char get();
-    std::string get_str(int length);
+    std::string get_str(size_t length);
     std::string get_line();
 
     char peek(size_t offset = 0);
-    std::string peek_str(int length);
+    std::string peek_str(size_t length);
 
     Locatable loc();
